+++ src/dictionary/DefaultMod.ts
@@ -2,7 +2,7 @@ import Data from "../data/Data"
 import DictionaryValue from "./DictionaryValue"
 import { observeType } from "./ObserveList"
 import TipValue, { TipValueInitOption } from "../lib/TipValue"
-import AttributeValue, { AttributeValueInitOption } from "../lib/AttributeValue"
+import { LocalValue, LocalValueInitOption, createLocalValue } from "../lib/AttrsValue"
 import InterfaceValue from "../lib/InterfaceValue"
 import { ArrayMapValueType } from "../lib/ArrayMap"
 
@@ -14,12 +14,7 @@ export interface DefaultModInitOption {
   $target?: string // 快捷格式化目标，内存指针指向对应的mod
   prop?: string
   name?: string
-  local?: {
-    parent?: AttributeValueInitOption
-    target?: AttributeValueInitOption
-    child?: AttributeValueInitOption
-    [prop: string]: undefined | AttributeValueInitOption
-  }
+  local?: LocalValueInitOption
   tip?: TipValueInitOption
   reactive?: {
     [prop: string]: undefined | reactiveFunction
@@ -38,12 +33,7 @@ class DefaultMod extends Data implements ArrayMapValueType {
   $prop: string
   $name: InterfaceValue<string>
   tip?: TipValue
-  $local?: {
-    parent?: AttributeValue
-    target?: AttributeValue
-    child?: AttributeValue
-    [prop: string]: undefined | AttributeValue
-  }
+  $local?: LocalValue
   $reactive?: {
     [prop: string]: undefined | reactiveFunction
   }
@@ -62,13 +52,7 @@ class DefaultMod extends Data implements ArrayMapValueType {
     if (initOption.tip !== undefined) {
       this.tip = new TipValue(initOption.tip)
     }
-    if (initOption.local) {
-      // 插件单独的设置，做特殊处理时使用，尽可能的将所有能用到的数据通过option做兼容处理避免问题
-      this.$local = {}
-      for (const prop in initOption.local) {
-        this.$local[prop] = new AttributeValue(initOption.local[prop])
-      }
-    }
+    this.$local = createLocalValue(initOption.local)
     this.$reactive = initOption.reactive
     this.$render = initOption.render
     this.$observe = initOption.observe
@@ -78,7 +62,7 @@ class DefaultMod extends Data implements ArrayMapValueType {
       return this.$render[prop]
     }
   }
-  $getAttribute(prop: string) {
+  $getLocalAttrs(prop: string) {
     if (this.$local) {
       return this.$local[prop]
     }