+++ src/lib/TipValue.ts
@@ -1,4 +1,4 @@
-import AttributeValue, { AttributeValueInitOption } from "./AttributeValue"
+import AttrsValue, { AttrsValueInitOption } from "./AttrsValue"
 
 export type locationType = 'top' | 'left' | 'right' | 'bottom' | 'topLeft' | 'topRight' | 'bottomLeft' | 'bottomRight' | 'leftTop' | 'leftBottom' | 'rightTop' | 'rightBottom'
 
@@ -6,7 +6,7 @@ export type TipValueInitOption = string | {
   data: string
   getData?: (...args: unknown[]) => string
   location?: locationType
-  local?: AttributeValueInitOption
+  attrs?: AttrsValueInitOption
 }
 
 class TipValue {
@@ -14,7 +14,7 @@ class TipValue {
   data?: string
   getData?: (...args: unknown[]) => string
   location: locationType
-  $local?: AttributeValue
+  $attrs?: AttrsValue
   constructor(initOption?: TipValueInitOption) {
     if (initOption === undefined || typeof initOption === 'string') {
       this.data = initOption
@@ -23,7 +23,7 @@ class TipValue {
       this.data = initOption.data
       this.getData = initOption.getData
       this.location = initOption.location || 'top'
-      this.$local = initOption.local ? new AttributeValue(initOption.local) : undefined
+      this.$attrs = initOption.attrs ? new AttrsValue(initOption.attrs) : undefined
     }
   }
 }